requirements.txt
python-telegram-bot
requests
os

